% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_amedas.R
\name{extract_timestamp}
\alias{extract_timestamp}
\title{Extracts download timestamp from file header}
\usage{
extract_timestamp(filename)
}
\arguments{
\item{filename}{}
}
\value{
timestamp as text
}
\description{
Extracts download timestamp from file header
}
