% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amedas_stations.R
\name{left_join_stations}
\alias{left_join_stations}
\title{Add station metadata to AMEDAS readings}
\usage{
left_join_stations(target)
}
\arguments{
\item{target}{AMADAS data frame to join to}
}
\value{
The data frame with station metadata added
}
\description{
Add station metadata to AMEDAS readings
}
