% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_measurements.R
\name{pivot_measurements}
\alias{pivot_measurements}
\title{Pivot the data frame so each measurement is it's own column.}
\usage{
pivot_measurements(data, min_quality, ignore_device_changes = FALSE)
}
\arguments{
\item{data}{A dataframe read by \code{read_amedas_csv}}

\item{min_quality}{The minimum acceptable level for 品質情報 (quality)}

\item{ignore_device_changes}{Silence warnings about changes in the "均質番号" field}
}
\value{
A pivoted \link[tibble:tibble-package]{tibble}
}
\description{
This requires specifying the acceptable quality (品質情報) level. All measurements
that fall below this threshold will be removed.
}
