% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_amedas.R
\name{read_amedas_csv}
\alias{read_amedas_csv}
\title{Read an AMEDAS CSV data file}
\usage{
read_amedas_csv(filename)
}
\arguments{
\item{filename}{File to read}
}
\value{
A \link[tibble:tibble-package]{tibble} with one row per time, station and observation type and
includes the measurement metadata (quality etc)
}
\description{
AMEDAS files downloaded from the JMA website can't be read with \code{read_csv}
due to multiple headers. \code{read_amedas_csv} attempts to parse the csv and
convert it into a tidy dataframe.
}
