% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{summarise_quality}
\alias{summarise_quality}
\title{Summarise record quality}
\usage{
summarise_quality(data)
}
\arguments{
\item{data}{Raw AMEDAS dataframe to summarise}
}
\value{
Tibble with count and proportion for each measurement
}
\description{
Summarise record quality
}
