% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_columns}
\alias{translate_columns}
\title{Translate column names to English}
\usage{
translate_columns(data)
}
\arguments{
\item{data}{AMEDAS dataframe to translate}
}
\value{
The same dataframe, but with the columns named in English
}
\description{
Convert column names to English. The measurement translations are currently incomplete.
}
