% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{visualize_quality_over_time}
\alias{visualize_quality_over_time}
\title{Visualise reading quality over time}
\usage{
visualize_quality_over_time(data)
}
\arguments{
\item{data}{AMEDAS dataframe to visualise}
}
\value{
ggplot graphic
}
\description{
Plot the reading quality values over time making it easy to asses the data
quality and decide \code{min_quality} values when pivoting
}
\examples{
raw_data <- read_amedas_csv("data.csv")
visualize_quality_over_time(raw_data)
# All < 8 readings are out of season readings, ok to drop them
data <- pivot_measurements(raw_data, min_quality = 8)
}
